pyproject.toml
[project]
name = "metal-bulletin"
version = "0.1.0"
description = "비철금속 시세 자동 수집 및 대시보드"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-genai>=1.74.0",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "rapidfuzz>=3.14.5",
    "requests>=2.33.1",
    "zstandard>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.15.1",
]

[tool.setuptools]
packages = ["scraper", "scraper.news", "parser", "parser.news", "exchange", "builder", "summarizer", "summarizer.providers"]

[tool.pytest.ini_options]
pythonpath = ["."]
